fixed goofyscript and usage of <any>, fixed react-query types, created custom hooks to get all records and delete a record byid. still need to edit server so records with foreign keys can be deleted without postgresql giving an error
